+++ app/vtselect/traces/tracecommon/tracecommon.go
@@ -19,11 +19,13 @@ var (
 	TraceMaxDurationWindow = flag.Duration("search.traceMaxDurationWindow", 1*time.Minute, "The window of searching for the rest trace spans after finding one span."+
 		"It allows extending the search start time and end time by -search.traceMaxDurationWindow to make sure all spans are included."+
 		"It affects both Jaeger's /api/traces and /api/traces/<trace_id> APIs.")
-	TraceServiceAndSpanNameLookbehind = flag.Duration("search.traceServiceAndSpanNameLookbehind", 3*24*time.Hour, "The time range of searching for service name and span name. "+
-		"It affects Jaeger's /api/services and /api/services/*/operations APIs.")
 	TraceSearchStep = flag.Duration("search.traceSearchStep", 24*time.Hour, "Splits the [0, now] time range into many small time ranges by -search.traceSearchStep "+
 		"when searching for spans by trace_id. Once it finds spans in a time range, it performs an additional search according to -search.traceMaxDurationWindow and then stops. "+
 		"It affects Jaeger's /api/traces/<trace_id> API.")
+	TraceFieldsLookbehind = flag.Duration("search.fieldsLookbehind", 2*time.Hour, "The default time range of searching for normal fields (tags and attributes)."+
+		"It affects various Tempo tag-related APIs.")
+	TraceStreamFieldsLookbehind = flag.Duration("search.streamFieldsLookbehind", 3*24*time.Hour, "The default time range of searching for stream fields (service name and span name)."+
+		"It affects Jaeger's /api/services and /api/services/*/operations APIs.")
 	TraceMaxTraces = flag.Int("search.maxTraces", 1000, "The maximum number of traces that can be returned in a single search request. "+
 		"Users may request with different limit value via query argument which shouldn't exceed this limit. This limit applies to Jaeger’s /api/traces API and Tempo's /api/search API.")
 	TraceMaxTags = flag.Int("search.maxTags", 1000, "The maximum number of tags (including service name, span name) that can be returned in a single search request. "+
@@ -34,6 +36,7 @@ var (
 	// deprecated flags. preserve here for backward compatibility. should be removed in future version.
 	_ = flag.Uint64("search.traceMaxServiceNameList", 1000, "Deprecated, see -search.maxTags.")
 	_ = flag.Uint64("search.traceMaxSpanNameList", 1000, "Deprecated, see -search.maxTags.")
+	_ = flag.Duration("search.traceServiceAndSpanNameLookbehind", 3*24*time.Hour, "Deprecated, see -search.tagsLookbehind.")
 )
 
 var (